week5

+++ app/src/main/java/com/example/hello/FirstActivity_1.java
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -16,10 +17,11 @@ public class FirstActivity_1 extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_first1);
+
         Intent intent=getIntent();
-        double dollar2=intent.getDoubleExtra("dollar_rate_key",0.00);
-        double euro2=intent.getDoubleExtra("euro_rate_key",0.00);
-        double won2=intent.getDoubleExtra("won_rate_key",0.00);
+        float dollar2=intent.getFloatExtra("dollar_rate_key",0.00f);
+        float euro2=intent.getFloatExtra("euro_rate_key",0.00f);
+        float won2=intent.getFloatExtra("won_rate_key",0.00f);
 
         dollar_rate=findViewById(R.id.dollar_rate);
         euro_rate=findViewById(R.id.euro_rate);
@@ -31,9 +33,9 @@ public class FirstActivity_1 extends AppCompatActivity {
     }
     public void save(View btn){
         Intent save=getIntent();
-        save.putExtra("dollar_rate_key2",Double.parseDouble(dollar_rate.getText().toString()));
-        save.putExtra("euro_rate_key2",Double.parseDouble(euro_rate.getText().toString()));
-        save.putExtra("won_rate_key2",Double.parseDouble(won_rate.getText().toString()));
+        save.putExtra("dollar_rate_key2",Float.parseFloat(dollar_rate.getText().toString()));
+        save.putExtra("euro_rate_key2",Float.parseFloat(euro_rate.getText().toString()));
+        save.putExtra("won_rate_key2",Float.parseFloat(won_rate.getText().toString()));
         setResult(1002,save);
         finish();
     }
